# services/scoring/openapi.yaml
openapi: 3.1.0
info:
  title: OCR Slip Gaji Scoring API
  description: 'Tahap terakhir pipeline slip gaji: memberi **skor keyakinan per field** — P(nilai ini
    benar), 0-1, terkalibrasi isotonic.


    Modelnya regresi logistik 9 variabel + one-hot field, dilatih pada 1.428 nilai dari 50 dokumen berlabel;
    out-of-fold AUC 0,826 · Gini 0,652 · ECE 0,037. Pada ambang 80: precision 93,4%, recall 75,6%.


    **Tidak ada keputusan di sini.** Skor mentah yang dikembalikan; pemetaan ke `confidence` 0/1 terjadi
    di kontrak `extract-ocr`, supaya ambangnya bisa diubah tanpa melatih ulang apa pun.


    **Pipeline asinkron:** service structuring POST /v1/scoring/jobs dan menerima 202; service ini memberi
    skor di latar, menyimpan hasilnya, dan mengirim callback berisi hasil akhir ke orchestrator. /v1/scoring/score
    melakukan pekerjaan yang sama secara sinkron. Semua endpoint kecuali /health memerlukan header X-API-Key.


    ## Conventions (the same for every nilam-ocr service)


    **Authentication.** Every endpoint except `/health`, `/ready` and `/metrics` requires the header `X-API-Key`.

    A missing or wrong key answers `401`. The services are reachable only from inside the cluster.


    **Envelope.** Every JSON response, success or error, has the same shape:

    `{status_code, status_desc, message, data, errors, request_id}`. `status_code` always equals the HTTP

    status. On success `errors` is null; on error `data` is null.
...
servers:
...
paths:
...
  /v1/scoring/jobs:
    post:
      tags:
      - Pipeline
      summary: Serahkan dokumen terstruktur ke tahap scoring (terakhir)
      description: '**Langkah 3 pipeline, asinkron, tahap terakhir. Dipanggil service structuring, bukan
        orchestrator.**


        Mencatat job (`scoring_jobs`, idempoten per request_id), menjawab **202 seketika**, lalu di latar:
        memberi skor keyakinan pada setiap field tiap slip, menyimpan hasilnya (`scoring_results`), dan
        mengirim callback `SCORING` yang membawa **hasil akhir** permintaan ini.


        Keluarannya skor 0-1 per field. Tidak ada skor tingkat dokumen dan tidak ada keputusan terima
        / tolak: ambang milik orchestrator.'
      operationId: submitScoringJob
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoringJobRequest'
        required: true
      responses:
        '202':
          description: The job was accepted (or already existed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
              examples:
                accepted:
                  summary: New job
                  value:
                    status_code: 202
                    status_desc: Accepted
                    message: Accepted
                    data:
                      request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                      stage: SCORING
                      status: PROCESSING
                      duplicate: false
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                duplicate:
                  summary: 'Same request_id sent again: nothing is re-run'
                  value:
                    status_code: 202
                    status_desc: Accepted
                    message: Accepted
                    data:
                      request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                      stage: SCORING
                      status: DONE
                      duplicate: true
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 401
                status_desc: Unauthorized
                message: Invalid or missing API key
                errors: Invalid or missing API key
                pipeline_last_stage: scoring
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 422
                status_desc: Unprocessable Entity
                message: 'body.request_id: Field required'
                errors: VALIDATION_ERROR
                pipeline_last_stage: scoring
      security:
      - APIKeyHeader: []
  /v1/scoring/jobs/{request_id}:
    get:
      tags:
      - Pipeline
      summary: Status and result of the scoring stage
      description: 'Status of this stage only and, once `DONE`, the two confidences plus the exact payload
        that was scored. Internal: the orchestrator Slip Gaji reads it (while it waits, and for its `GET
        /v1/extract-ocr/{request_id}`); the central orchestrator receives the final result in the callback.
        Also useful to audit a score.'
      operationId: getScoringJob
      security:
      - APIKeyHeader: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          title: Request Id
      responses:
        '200':
          description: The job exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoringJobStatusResponse'
              examples:
                done:
                  summary: Finished
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                      stage: SCORING
                      created_at: '2026-09-18T04:00:01+00:00'
                      status: DONE
                      result:
                        slips:
                        - slip_no: 1
                          scores:
                            nama_karyawan: 76.71
                            jabatan: 76.71
                            gaji_pokok: 93.6
                            tunjangan_makan: 88.89
                            total_pendapatan: 92.47
                            gaji_bersih: 92.47
                        threshold: 80.0
                        model: conf_v2
                      updated_at: '2026-09-18T04:00:01+00:00'
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                failed:
                  summary: Failed
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                      stage: SCORING
                      created_at: '2026-09-18T04:00:01+00:00'
                      status: FAILED
                      error_message: 'Unsupported document_type: ktp. Supported: [''slip_gaji'']'
                      updated_at: '2026-09-18T04:00:01+00:00'
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              example:
                status_code: 401
                status_desc: Unauthorized
                message: Invalid or missing API key
                errors: Invalid or missing API key
                pipeline_last_stage: scoring
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No scoring job for this request_id
          content:
            application/json:
              example:
                status_code: 404
                status_desc: Not Found
                message: 'No SCORING job found for request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0'
                errors: 'No SCORING job found for request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0'
                request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                pipeline_last_stage: scoring
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              example:
                status_code: 422
                status_desc: Unprocessable Entity
                message: 'path.request_id: Field required'
                errors: VALIDATION_ERROR
                request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                pipeline_last_stage: scoring
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/scoring/outbox:
    get:
      tags:
      - Pipeline
      summary: Backlog of this stage's undelivered callbacks and hand-offs
      description: 'Operational view of the transactional outbox (`PIPELINE_OUTBOX`): how many callbacks
        and hand-offs of this stage are still to be delivered, how old the oldest one is, and how many
        were given up on (dead letters). No other service watches the `pipeline_outbox` table, so this
        is where a monitor should look; the relay also logs a warning when the backlog is older than `PIPELINE_OUTBOX_STALE_AFTER_SECONDS`
        or a dead letter is waiting.'
      operationId: getScoringOutboxStatus
      responses:
        '200':
          description: The backlog of this stage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutboxStatusResponse'
              examples:
                idle:
                  summary: Nothing waiting
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      enabled: true
                      stage: SCORING
                      pending: 0
                      retrying: 0
                      dead_letters: 0
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                backlog:
                  summary: The orchestrator has been down for a while
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      enabled: true
                      stage: SCORING
                      pending: 12
                      retrying: 12
                      oldest_pending_seconds: 754.3
                      dead_letters: 1
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                disabled:
                  summary: '`PIPELINE_OUTBOX` is off'
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      enabled: false
                      stage: SCORING
                      pending: 0
                      retrying: 0
                      dead_letters: 0
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 401
                status_desc: Unauthorized
                message: Invalid or missing API key
                errors: Invalid or missing API key
                pipeline_last_stage: scoring
      security:
      - APIKeyHeader: []
  /v1/scoring/outbox/release:
    post:
      tags:
      - Pipeline
      summary: Send this stage's dead letters again
      description: 'Puts the dead letters of this stage back in the outbox queue, due immediately, so
        the relay delivers them again: use it after the receiver was fixed (for example the orchestrator''s
        callback contract, or a next stage that was answering 4xx). Optional query `request_id` limits
        it to one request. A message that fails again goes through the same retry and dead-letter rules.
        `409` when `PIPELINE_OUTBOX` is off on this service.'
      operationId: releaseScoringOutbox
      security:
      - APIKeyHeader: []
      parameters:
      - name: request_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
      responses:
        '200':
          description: Dead letters released
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutboxReleaseResponse'
              examples:
                released:
                  summary: One request
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      stage: SCORING
                      request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                      released: 1
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                nothing:
                  summary: Nothing to release
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      stage: SCORING
                      released: 0
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              example:
                status_code: 401
                status_desc: Unauthorized
                message: Invalid or missing API key
                errors: Invalid or missing API key
                pipeline_last_stage: scoring
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The outbox is off on this service
          content:
            application/json:
              example:
                status_code: 409
                status_desc: Conflict
                message: 'PIPELINE_OUTBOX is off: there is no outbox to release'
                errors: 'PIPELINE_OUTBOX is off: there is no outbox to release'
                pipeline_last_stage: scoring
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              example:
                status_code: 422
                status_desc: Unprocessable Entity
                message: 'query.request_id: Input should be a valid string'
                errors: VALIDATION_ERROR
                pipeline_last_stage: scoring
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          examples:
          - 400
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          examples:
          - Bad Request
        message:
          type: string
          title: Message
          description: Human-readable cause; safe to log
          examples:
          - Uploaded file is empty
        data:
          type: 'null'
          title: Data
          description: Always null on error
        errors:
          type: string
          title: Errors
          description: Machine-readable cause. `VALIDATION_ERROR` for 422 (branch on this, the message
            names a different field every time); for every other status it repeats `message`
          examples:
          - Uploaded file is empty
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: Present when the error is tied to a known request_id
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
      type: object
      required:
      - status_code
      - status_desc
      - message
      - errors
      title: ErrorResponse
      description: Every error response.
...
    GuardrailsPayload:
      properties:
        passed:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Passed
          description: 'true: dokumen boleh lanjut ke structuring'
          examples:
          - true
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
          description: Alasan ditahan; null bila lolos
          examples:
          - null
        verdict:
          anyOf:
          - type: string
          - type: 'null'
          title: Verdict
          description: Putusan gabungan
          examples:
          - slip_gaji
        skipped:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Skipped
          description: Guardrail yang dimatikan (`GUARDRAIL_<NAMA>_ENABLED=false`)
          examples:
          - []
        unavailable:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Unavailable
          description: Guardrail yang menyala tetapi tidak menjawab (dilewati, fail-open)
          examples:
          - []
        skipped_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Skipped Reason
          description: Kenapa tidak ada putusan
        document:
          anyOf:
          - $ref: '#/components/schemas/GuardrailsDocument'
          - type: 'null'
      additionalProperties: true
      type: object
      title: GuardrailsPayload
      description: Laporan gabungan ketiga guardrail (guardrail-blank, guardrail-blur, guardrail-identity),
        dihasilkan tahap OCR karena modelnya membaca teks OCR, lalu diteruskan tanpa diubah ke structuring
        dan scoring.
...
    JobAccepted:
      properties:
        request_id:
          type: string
          title: Request Id
          description: Echo of the submitted request_id
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        stage:
          type: string
          enum:
          - OCR
          - STRUCTURING
          - SCORING
          title: Stage
          description: Pipeline stage that accepted the job
          examples:
          - OCR
        status:
          type: string
          enum:
          - PROCESSING
          - DONE
          - FAILED
          title: Status
          description: '`PROCESSING` for a newly accepted job. When `duplicate` is true: the current status
            of the job that already exists for this request_id'
          examples:
          - PROCESSING
        duplicate:
          type: boolean
          title: Duplicate
          description: 'true: this request_id was already accepted before and the work is NOT run again
            (idempotent). Exception: a job whose status is `FAILED` is re-claimed and run again, and answers
            `duplicate: false`'
          default: false
          examples:
          - false
      type: object
      required:
      - request_id
      - stage
      - status
      title: JobAccepted
      description: '`data` of `POST /v1/<stage>/jobs`.'
    JobAcceptedResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          default: 202
          examples:
          - 202
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          default: Accepted
          examples:
          - Accepted
        message:
          type: string
          title: Message
          description: Human-readable outcome
          default: Accepted
          examples:
          - Accepted
        errors:
          type: 'null'
          title: Errors
          description: Always null on success
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: 'The request_id this response belongs to: the one sent in the form / body / path,
            otherwise the `X-Request-ID` header, otherwise one generated by the service'
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        data:
          $ref: '#/components/schemas/JobAccepted'
      type: object
      required:
      - data
      title: JobAcceptedResponse
      description: Envelope of `POST /v1/<stage>/jobs`.
    OutboxRelease:
      properties:
        stage:
          type: string
          enum:
          - OCR
          - STRUCTURING
          - SCORING
          title: Stage
          description: The stage whose dead letters were released
          examples:
          - OCR
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: Only this request's dead letters were released; null = every dead letter of the
            stage
        released:
          type: integer
          title: Released
          description: Dead letters put back in the queue, due immediately
          examples:
          - 1
      type: object
      required:
      - stage
      - released
      title: OutboxRelease
      description: The `data` of `POST /v1/<stage>/outbox/release`.
    OutboxReleaseResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          default: 200
          examples:
          - 200
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          default: OK
          examples:
          - OK
        message:
          type: string
          title: Message
          description: Human-readable outcome
          default: Success
          examples:
          - Success
        errors:
          type: 'null'
          title: Errors
          description: Always null on success
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: 'The request_id this response belongs to: the one sent in the form / body / path,
            otherwise the `X-Request-ID` header, otherwise one generated by the service'
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        data:
          $ref: '#/components/schemas/OutboxRelease'
      type: object
      required:
      - data
      title: OutboxReleaseResponse
      description: Envelope of `POST /v1/<stage>/outbox/release`.
...
    OutboxStatusResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          default: 200
          examples:
          - 200
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          default: OK
          examples:
          - OK
        message:
          type: string
          title: Message
          description: Human-readable outcome
          default: Success
          examples:
          - Success
        errors:
          type: 'null'
          title: Errors
          description: Always null on success
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: 'The request_id this response belongs to: the one sent in the form / body / path,
            otherwise the `X-Request-ID` header, otherwise one generated by the service'
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        data:
          $ref: '#/components/schemas/OutboxStatus'
      type: object
      required:
      - data
      title: OutboxStatusResponse
      description: Envelope of `GET /v1/<stage>/outbox`.
...
    ScoringJobRequest:
      properties:
        request_id:
          type: string
          title: Request Id
          examples:
          - OCR_9cb01af2-493d-446d-b191-af120333f6d0
        document_type:
          type: string
          title: Document Type
          default: slip_gaji
          examples:
          - slip_gaji
        guardrails:
          anyOf:
          - $ref: '#/components/schemas/GuardrailsPayload'
          - type: 'null'
          description: Laporan guardrail dari tahap OCR; dikembalikan apa adanya di hasil akhir
        ocr:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Ocr
          description: Hasil tahap OCR; opsional
        structuring:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Structuring
          description: 'Hasil tahap structuring. Boleh dihilangkan bila `PIPELINE_HANDOFF_BY_REFERENCE`
            menyala: service ini membacanya dari `structuring_results` memakai request_id'
        pipeline_name_sequence:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Pipeline Name Sequence
          description: Urutan service permintaan ini; null = pipeline penuh
          examples:
          - null
        column_confidence_threshold:
          anyOf:
          - additionalProperties:
              type: number
            type: object
          - type: 'null'
          title: Column Confidence Threshold
          description: Ambang per field (0-1) dari permintaan; disimpan bersama job supaya GET menjawab
            sama
          examples:
          - gaji_bersih: 0.9
      type: object
      required:
      - request_id
      title: ScoringJobRequest
...
    ScoringJobStatusResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          default: 200
          examples:
          - 200
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          default: OK
          examples:
          - OK
        message:
          type: string
          title: Message
          description: Human-readable outcome
          default: Success
          examples:
          - Success
        errors:
          type: 'null'
          title: Errors
          description: Always null on success
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: 'The request_id this response belongs to: the one sent in the form / body / path,
            otherwise the `X-Request-ID` header, otherwise one generated by the service'
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        data:
          $ref: '#/components/schemas/ScoringJobStatus'
      type: object
      required:
      - data
      title: ScoringJobStatusResponse
...
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      description: API key for this service
      in: header
      name: X-API-Key
...